+++ sdk/cosmos/azure_data_cosmos/tests/multi_write.rs
@@ -1,0 +1,5 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+#![cfg(test_category = "multi_write")]
+
+mod multi_write_tests;