+++ sdk/cosmos/azure_data_cosmos/tests/multi_write_tests/mod.rs
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+mod cosmos_multi_write;
+
+#[path = "../framework/mod.rs"]
+mod framework;